Move base64 to src/common
