+++ src/runtime/keymgmt/webcrypto.ts
@@ -12,10 +12,10 @@ import idb, {ObjectStore} from 'idb';
 import rs from 'jsrsasign';
 
 import {crypto} from '../../platform/crypto-web.js';
-import {decode, encode} from './base64';
-import {DeviceKey, Key, PrivateKey, PublicKey, RecoveryKey, SessionKey, WrappedKey} from './keys';
-import {KeyGenerator, KeyStorage} from './manager';
-import {TestableKey} from './testing/cryptotestutils';
+import {decode, encode} from '../../common/io/base64.js';
+import {DeviceKey, Key, PrivateKey, PublicKey, RecoveryKey, SessionKey, WrappedKey} from './keys.js';
+import {KeyGenerator, KeyStorage} from './manager.js';
+import {TestableKey} from './testing/cryptotestutils.js';
 
 const DEVICE_KEY_ALGORITHM = 'RSA-OAEP';
 const X509_CERTIFICATE_ALGORITHM = 'RSA-OAEP';
